organize directory structure a bit
