+++ module-karel/src/runs.ts
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import {Submission} from 'dt-types'
 import {Partitions} from './partitions'
-import {HwConfig} from './homework'
+import {HwConfig} from './types/homework'
 import path from 'path'
 import {StudentList} from "classroom-api";
 import {config} from "./config";
