+++ module-karel/src/scripts/notifySum.ts
@@ -1,10 +1,10 @@
 import {sendEmails} from "classroom-api";
 import {convertToCsv, summarizeResults} from "./sumResults";
-import {summaries} from "./templates";
+import {summaries} from "../templates";
 import { ArgumentParser } from 'argparse'
 import fs from "fs";
 import path from "path";
-import {data_path} from "./runs";
+import {data_path} from "../runs";
 
 
 function notifyEmails() {
